Update to jsbundling-rails
