+++ .gitignore
@@ -32,3 +32,8 @@ config/environments/*.local.yml
 /.ruby-version
 /.byebug_history
 /spec/examples.txt
+
+/app/assets/builds/*
+!/app/assets/builds/.keep
+
+/node_modules
