+++ README.md
@@ -29,7 +29,7 @@ To start Solr, you can use the `solr_wrapper` command. However, if starting from
 
 Start the rails app
 
-    $ rails s
+    $ bin/dev
 
 ## Getting data in development
 
@@ -37,9 +37,9 @@ For fixture data, you can run `solr_wrapper` locally as above, and populate samp
 
 ## "Logging in" as a User in development
 
-Given that this app is using shibboleth + devise for login, it can be tricky to get a user context set.  The user itself can be set by starting rails with a `REMOTE_USER` env var set (e.g. `REMOTE_USER=jstanford rails s`).
+Given that this app is using shibboleth + devise for login, it can be tricky to get a user context set.  The user itself can be set by starting rails with a `REMOTE_USER` env var set (e.g. `REMOTE_USER=jstanford bin/dev`).
 
-If you need to set your affiliation attribute (e.g. `stanford:staff`) for things like testing article search, you can start rails with a `suAffiliation` env var set (e.g. `REMOTE_USER=jstanford suAffiliation=stanford:staff rails s`).
+If you need to set your affiliation attribute (e.g. `stanford:staff`) for things like testing article search, you can start rails with a `suAffiliation` env var set (e.g. `REMOTE_USER=jstanford suAffiliation=stanford:staff bin/dev`).
 
 One caveat to this is that you may still need to go through the login path `/sso/login` in order to ensure proper session setup (e.g. eds guest flag set).
