+++ app/assets/config/manifest.js
@@ -3,3 +3,4 @@
 //= link application.js
 //= link application.css
 //= link print.css
+//= link_tree ../builds
