+++ app/javascript/analytics.js
