+++ app/javascript/article.js
