+++ app/javascript/blacklight_hierarchy.js
