+++ app/javascript/bootstrap-modal-addon.js
