+++ app/javascript/course_reserves.js
