+++ app/javascript/facet-options-checkboxes.js
