+++ app/javascript/home_page_facet_collapse.js
