+++ app/javascript/jquery.accordion-section.js
