+++ app/javascript/jquery.clear-input-text.js
