+++ app/javascript/jquery.libraryh3lp.js
