+++ app/javascript/jquery.live-lookup.js
