+++ app/javascript/jquery.long-lists.js
