+++ app/javascript/jquery.managed-purl.js
