+++ app/javascript/jquery.plug-google-content.js
