+++ app/javascript/jquery.side-nav-minimap.js
