+++ app/javascript/jquery.stackmap.js
