+++ app/javascript/jquery.turbolinks-cursor.js
