+++ app/javascript/location-hours.js
