+++ app/javascript/select-all.js
