+++ app/javascript/sfx-panel.js
