+++ app/javascript/truncate.js
