+++ app/javascript/update-hidden-inputs-by-checkbox.js
